requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Mail==0.9.1
Werkzeug==2.3.7
email-validator==2.0.0.post2
python-dotenv==1.0.0
Flask-Migrate==4.0.5
Pillow==10.0.0
Flask-Bcrypt==1.0.1
pandas==2.0.3
openpyxl==3.1.2
requests==2.31.0
itsdangerous==2.1.2
cloudinary>=1.40.0
APScheduler==3.10.4
psycopg[binary]==3.1.12
gunicorn==21.2.0
